requirements.txt
langchain
langchain-openai
langchain-community
langchain-huggingface
langchain-text-splitters
langchain-chroma
langchain-core==0.3.65

langchain_groq
langserve[all]

# OpenAI & Tokenization
openai
tiktoken

# Embeddings & Vector Search
sentence-transformers
torch==2.3.0
faiss-cpu
numpy==1.26.4

# Vector DB
chromadb

# Document Parsing
pypdf
pymupdf

# Web Data Loaders
wikipedia
bs4
arxiv
beautifulsoup4

# Env & Notebook
python-dotenv
ipykernel
rich

# API & UI
fastapi
uvicorn
streamlit

# Optional
httpx>=0.27.0

arxiv
wikipedia
duckduckgo-search

#database
mysql-connector-python

validators==0.28.1
youtube_transcript_api
